% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{computeConfusionMatrix}
\alias{computeConfusionMatrix}
\title{Compute confusion matrix}
\usage{
computeConfusionMatrix(signals, simulationSettings)
}
\arguments{
\item{signals}{An object created using \code{\link[=runDiscoverySystem]{runDiscoverySystem()}}.}

\item{simulationSettings}{An object created using \code{\link[=createSimulationSettings]{createSimulationSettings()}}.}
}
\value{
A tibble.
}
\description{
Compute confusion matrix
}
