% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationSettings.R
\name{computeNtarget}
\alias{computeNtarget}
\title{Compute number of subjects in target population at a specific look.}
\usage{
computeNtarget(
  lookId,
  nTargetFirstLook,
  nTargetNextLooks,
  nTargetNextLooksDelta
)
}
\arguments{
\item{lookId}{The look at which to compute the number of subjects.}

\item{nTargetFirstLook}{Number of subjects in the target population at the start (look 1).}

\item{nTargetNextLooks}{Number of subjects observed in subsequent looks.}

\item{nTargetNextLooksDelta}{Change to the target delta at each look.}
}
\value{
The number of subjects in the target population at look \code{lookId}.
}
\description{
Compute number of subjects in target population at a specific look.
}
