% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{createDatabaseSettings}
\alias{createDatabaseSettings}
\title{Create database settings}
\usage{
createDatabaseSettings(sampleSizeMultiplier = 1)
}
\arguments{
\item{sampleSizeMultiplier}{The relative sample size of the database.}
}
\value{
A settings object
}
\description{
Create database settings
}
