% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDiscovery.R
\name{createDiscoverySystemSettings}
\alias{createDiscoverySystemSettings}
\title{Create setting for the discovery system}
\usage{
createDiscoverySystemSettings(alpha = 0.05, useOracleForCv = FALSE)
}
\arguments{
\item{alpha}{The family-wise type 1 error rate we're aiming for.}

\item{useOracleForCv}{Compute critical values on the actual sample sizes? If FALSE,
use expected sample sizes instead.}
}
\value{
A settings object
}
\description{
Create setting for the discovery system
}
