% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{createExposureOutcomeSettings}
\alias{createExposureOutcomeSettings}
\title{Create exposure-outcome settings}
\usage{
createExposureOutcomeSettings(
  nTarget = 100,
  nComparator = nTarget * 10,
  backgroundRate = 0.001,
  logRrMean = 0,
  logRrSd = 0,
  riskStart = 0,
  riskEnd = 21
)
}
\arguments{
\item{nTarget}{Number of subjects in the target population.}

\item{nComparator}{Number of subjects in the comparator (counterfactual) population.}

\item{backgroundRate}{Poisson background rate of the outcome.}

\item{logRrMean}{The mean of the log relative distribution across databases.}

\item{logRrSd}{The standard deviation (SD) of the log relative distribution across databases.}

\item{riskStart}{Start of the true risk window (relative to exposure start).}

\item{riskEnd}{End of the true risk window (relative to exposure start).}
}
\value{
A settings object
}
\description{
Create exposure-outcome settings
}
\details{
For each database, an effect size is sampled from the relative risk distribution. The relative risk
indicates the increase in the risk during the risk window.
}
