% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{createMethodSettings}
\alias{createMethodSettings}
\title{Create method settings}
\usage{
createMethodSettings(systematicErrorMean = 0, systematicErrorSd = 0)
}
\arguments{
\item{systematicErrorMean}{The mean of the systematic error distribution.}

\item{systematicErrorSd}{The standard deviation (SD) of the systematic error distribution.}
}
\value{
A settings object
}
\description{
Create method settings
}
