% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationSettings.R
\name{createSimulationSettings}
\alias{createSimulationSettings}
\title{Create simulation settings}
\usage{
createSimulationSettings(
  exposureOutcomeSettings = c(lapply(rep(1000, 90), createExposureOutcomeSettings,
    logRrMean = 0, logRrSd = 0), lapply(rep(1000, 10), createExposureOutcomeSettings,
    logRrMean = log(2), logRrSd = 0.25)),
  timeAtRiskSettings = list(createTimeAtRiskSettings(0, 7), createTimeAtRiskSettings(0,
    21), createTimeAtRiskSettings(0, 42), createTimeAtRiskSettings(0, 90)),
  methodSettings = list(createMethodSettings(0.05, 0.05), createMethodSettings(0.1,
    0.1), createMethodSettings(0.2, 0.2)),
  databaseSettings = list(createDatabaseSettings(1), createDatabaseSettings(0.5),
    createDatabaseSettings(2)),
  looks = 10
)
}
\arguments{
\item{exposureOutcomeSettings}{A list of objects created using \code{\link[=createExposureOutcomeSettings]{createExposureOutcomeSettings()}}.}

\item{timeAtRiskSettings}{A list of objects created using \code{\link[=createTimeAtRiskSettings]{createTimeAtRiskSettings()}}.}

\item{methodSettings}{A list of objects created using \code{\link[=createMethodSettings]{createMethodSettings()}}.}

\item{databaseSettings}{A list of objects created using \code{\link[=createDatabaseSettings]{createDatabaseSettings()}}.}

\item{looks}{The number of looks over time.}
}
\value{
A settings object
}
\description{
Create simulation settings
}
