% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{createTimeAtRiskSettings}
\alias{createTimeAtRiskSettings}
\title{Create time-at-risk settings}
\usage{
createTimeAtRiskSettings(start = 0, end = 21)
}
\arguments{
\item{start}{The start of the time-at-risk (relative to exposure start).}

\item{end}{The end of the time-at-risk (relative to exposure start).}
}
\value{
A settings object
}
\description{
Create time-at-risk settings
}
