% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverySystemSimulator.R
\docType{package}
\name{DiscoverySystemSimulator-package}
\alias{DiscoverySystemSimulator}
\alias{DiscoverySystemSimulator-package}
\title{DiscoverySystemSimulator: Simulating Effect discovery in an Active Surveillance System}
\description{
An R package that simulates effect discovery in an active surveillance system.
}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org} (\href{https://orcid.org/0000-0002-0817-5361}{ORCID})

}
\keyword{internal}
