% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{plotDecisionCurves}
\alias{plotDecisionCurves}
\title{Plot utility}
\usage{
plotDecisionCurves(evaluation, labels = NULL, alphas = NULL, fileName = NULL)
}
\arguments{
\item{evaluation}{An object generated by \code{\link[=evaluateIterations]{evaluateIterations()}}.}

\item{labels}{Which labels to plot? (Values of the \code{label} column). If \code{NULL} then all values are plotted.}

\item{alphas}{Which alphas to plot? (Values of the \code{alpha} column). If \code{NULL} then all alphas are plotted.}

\item{fileName}{Optional: the name of the file to save the plot to.}
}
\value{
A GGPlot object.
}
\description{
Plot utility
}
