% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{plotFalsePositiveNegatives}
\alias{plotFalsePositiveNegatives}
\title{Plot distribution of false positives and negatives}
\usage{
plotFalsePositiveNegatives(
  confusionMatrices,
  labels = NULL,
  facet = NULL,
  cumulative = TRUE,
  alpha = 0.05,
  fileName = NULL
)
}
\arguments{
\item{confusionMatrices}{An object generated by \code{\link[=evaluateIterations]{evaluateIterations()}}.}

\item{labels}{Which labels to plot? (Values of the \code{label} column). If \code{NULL} then all values are plotted.}

\item{cumulative}{Should cumulative probabilities be shown?}

\item{alpha}{The nominal alpha to plot. NA for no alpha.}

\item{fileName}{Optional: the name of the file to save the plot to.}
}
\value{
A GGPlot object.
}
\description{
Plot distribution of false positives and negatives
}
