% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDiscovery.R
\name{runDiscoverySystem}
\alias{runDiscoverySystem}
\title{Run discovery system}
\usage{
runDiscoverySystem(
  simulation = simulateDiscoverySystem(),
  discoverySystemSettings = createDiscoverySystemSettings()
)
}
\arguments{
\item{simulation}{An object generated by \code{\link[=simulateDiscoverySystem]{simulateDiscoverySystem()}}.}

\item{discoverySystemSettings}{An object generated by \code{\link[=createDiscoverySystemSettings]{createDiscoverySystemSettings()}}.}
}
\value{
Returns a tibble with signal flags for each row in the simulation.
}
\description{
Run discovery system
}
