% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleIterations.R
\name{runSimulationIterations}
\alias{runSimulationIterations}
\title{Run multiple iterations of the same simulation.}
\usage{
runSimulationIterations(
  simulationsFolder,
  simulationSettings = createSimulationSettings(),
  iterations = 100,
  threads = 4
)
}
\arguments{
\item{simulationsFolder}{The folder where the simulation objects should be written.}

\item{simulationSettings}{An object created using \code{\link[=createSimulationSettings]{createSimulationSettings()}}.}

\item{iterations}{The number of iterations.}

\item{threads}{The number of threads to use. Ideally, this number should
be lower than the number of CPU cores.}
}
\value{
This function does not return anything, but writes all output to the specified
folder. This folder will contain one file per iteration.
}
\description{
Run multiple iterations of the same simulation.
}
