% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.R
\name{simulateDiscoverySystem}
\alias{simulateDiscoverySystem}
\title{Simulate estimates generated by a discovery system.}
\usage{
simulateDiscoverySystem(simulationSettings = createSimulationSettings())
}
\arguments{
\item{simulationSettings}{An object created using \code{\link[=createSimulationSettings]{createSimulationSettings()}}.}
}
\value{
A tibble with counts and estimates per exposure-outcome, database, method, time-at-risk, and
look at the data. A \code{profiles} attribute contains the likelihood profiles.
}
\description{
Simulate estimates generated by a discovery system.
}
