% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{computeConfusionMatrix}
\alias{computeConfusionMatrix}
\title{Compute confusion matrix}
\usage{
computeConfusionMatrix(signals, simulationSettings, level = "exposure-outcome")
}
\arguments{
\item{signals}{An object created using \code{\link[=runDiscoverySystem]{runDiscoverySystem()}}.}

\item{simulationSettings}{An object created using \code{\link[=createSimulationSettings]{createSimulationSettings()}}.}

\item{level}{The level at which to compute the confusion matrix. Currently
supports "exposure-outcome" and "across looks".}
}
\value{
A tibble.
}
\description{
Compute confusion matrix
}
