% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{createDatabaseSettings}
\alias{createDatabaseSettings}
\title{Create method settings}
\usage{
createDatabaseSettings(sampleSizeMultiplier = 1)
}
\arguments{
\item{systematicErrorMean}{The mean of the systematic error distribution.}

\item{systematicErrorDSd}{The standard deviation (SD) of the systematic error distribution.}
}
\value{
A settings object
}
\description{
Create method settings
}
