% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleIterations.R
\name{evaluateIterations}
\alias{evaluateIterations}
\title{Evaluate iterations}
\usage{
evaluateIterations(signalsFolder)
}
\arguments{
\item{signalsFolder}{The folder containing the signal objects.}
}
\value{
A tibble combining the confusion matrices of all iterations.
}
\description{
Evaluate iterations
}
\details{
Requires \code{\link[=runSimulationIterations]{runSimulationIterations()}} to have already been executed.
}
