% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDiscovery.R
\name{runDiscoverySystem}
\alias{runDiscoverySystem}
\title{Run discovery system}
\usage{
runDiscoverySystem(
  simulation = simulateDiscoverySystem(),
  discoverySystemSettings = createDiscoverySystemSettings(),
  cacheFolder = NULL,
  cvCacheFile = NULL
)
}
\arguments{
\item{simulation}{An object generated by \code{\link[=simulateDiscoverySystem]{simulateDiscoverySystem()}}.}

\item{discoverySystemSettings}{An object generated by \code{\link[=createDiscoverySystemSettings]{createDiscoverySystemSettings()}}.}

\item{cacheFolder}{Optionally, a folder where intermediate artifacts (e.g. fitted
systematic error distributions) can be cached. If \code{NULL}, no
caching is performed.}

\item{cvCacheFile}{Optionally, the name of a global cache file for critical values.
If \code{NULL}, no caching is performed.}
}
\value{
Returns a tibble with signal flags for each row in the simulation.
}
\description{
Run discovery system
}
