% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleIterations.R
\name{runDiscoverySystemIterations}
\alias{runDiscoverySystemIterations}
\title{Run signal discovery on multiple iterations of the same simulation.}
\usage{
runDiscoverySystemIterations(
  simulationsFolder,
  signalsFolder,
  discoverySystemSettings = createDiscoverySystemSettings(),
  threads = 4,
  useCaching = TRUE,
  cvCacheFile = NULL
)
}
\arguments{
\item{simulationsFolder}{The folder where the simulation objects were written.}

\item{signalsFolder}{The folder where the signal objects will be written.}

\item{discoverySystemSettings}{An object generated by \code{\link[=createDiscoverySystemSettings]{createDiscoverySystemSettings()}}.}

\item{threads}{The number of threads to use. Ideally, this number should
be lower than the number of CPU cores.}

\item{useCaching}{Cache intermediate artifacts in the \code{simulationsFolder}? This will
greatly speed up subsequent runs.}

\item{cvCacheFile}{Optionally, the name of a global cache file for critical values.
If \code{NULL}, no caching is performed.}
}
\value{
This function does not return anything, but writes all output to the specified
folder. This folder will contain one file per iteration.
}
\description{
Run signal discovery on multiple iterations of the same simulation.
}
\details{
Requires \code{\link[=runSimulationIterations]{runSimulationIterations()}} to have already been executed.
}
